+++ src/main/java/se/lexicom/jpa_assignement/DAO/RecipeIngredientDAOImpl.java
@@ -1,4 +1,4 @@
-package se.lexicom.jpa_assignement.repository;
+package se.lexicom.jpa_assignement.DAO;
 
 import org.springframework.stereotype.Repository;
 import se.lexicom.jpa_assignement.exceptions.ExceptionManager;
@@ -6,8 +6,7 @@ import se.lexicom.jpa_assignement.model.RecipeIngredient;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
-import javax.persistence.Query;
-import java.util.Collection;
+import java.util.List;
 
 @Repository
 public class RecipeIngredientDAOImpl implements RecipeIngredientDAO {
@@ -25,23 +24,32 @@ public class RecipeIngredientDAOImpl implements RecipeIngredientDAO {
     }
 
     @Override
-    public RecipeIngredient delete(RecipeIngredient recipeIngredient) {
+    public RecipeIngredient delete(RecipeIngredient recipeIngredient) throws ExceptionManager {
+        if (recipeIngredient == null) {
+            throw new ExceptionManager("Can not delete item: " + recipeIngredient);
+        }
         entityManager.remove(recipeIngredient);
         return recipeIngredient;
     }
 
     @Override
-    public Collection<RecipeIngredient> findAll() {
-        return entityManager.createQuery("SELECT r FROM RecipeIngredient r").getResultList();
+    public List<RecipeIngredient> findAll() {
+        return entityManager.createQuery("SELECT r FROM RecipeIngredient r", RecipeIngredient.class).getResultList();
     }
 
     @Override
-    public RecipeIngredient findById(String id) {
-        return entityManager.find(RecipeIngredient.class, id);
+    public RecipeIngredient findById(String recipeIngredientId) throws ExceptionManager {
+        if (recipeIngredientId == null) {
+            throw new ExceptionManager("Can not find item: " + recipeIngredientId);
+        }
+        return entityManager.find(RecipeIngredient.class, recipeIngredientId);
     }
 
     @Override
-    public RecipeIngredient update(RecipeIngredient recipeIngredient) {
+    public RecipeIngredient update(RecipeIngredient recipeIngredient) throws ExceptionManager {
+        if (recipeIngredient == null) {
+            throw new ExceptionManager("Can not update item: " + recipeIngredient);
+        }
         return entityManager.merge(recipeIngredient);
     }
